Properly future-proof GPT Header
calculate_crc takes the rest of the bytes in a block, since header_size may be bigger than 92.
We can't write it back out, but we should be able to validate it's CRC.

This also fixes an existing bug in calculate_crc, source_bytes went up to header.header_size, but is only RawHeader bytes.